Update README.md
Welcome to our online store! Our website is created using a mix of HTML, CSS, JavaScript, and PHP, all working together to provide you with a great shopping experience. You can easily browse through a wide variety of products, add the ones you like to your shopping cart, and then smoothly buy them.

We've built this website to make your shopping journey simple. You can create an account, log in, and manage your shopping cart. When you're ready to buy, we've made the checkout process as easy as possible.

Behind the scenes, there's an area for administrators to handle products and keep things organized. We've also taken care to store all the important information about the products and users in a safe place.

The project is neatly organized. Different types of files are placed in specific folders, making it easier for us to manage. If you want to try it out, we've provided simple steps to set it up on your computer.

Our goal is to offer a user-friendly, secure, and interactive online shopping experience. You're more than welcome to contribute to our project, and everything here is under the MIT License, which means we encourage collaboration and improvement.

Come and explore our eCommerce Full Stack Development Project to enjoy a smooth shopping experience! If you have any questions or need help, please don't hesitate to get in touch using the contact details provided.